Feature/remove compiler warnings (#210)
* latest from coda-oss

* Update .gitignore

* use Standard C types "int8_t", etc. instead of our own typedefs

* use standard C types uint8_t, etc. instead of nitf_Uint8

* replace nitf::Uint8 etc with int8_t

* remove a bunch of compiler warnings (copied from develop/jdsmith)

* use std::to_string() instead of str::toString()

* allow unit-tests to work with Visual Studio/GTest

* rename Reader.c and Writer.c to avoid modules with the same name (and VS warnings)

* WIN32 isn't always defined, _WIN32 is (leading underscore)

* min/max should not be #define'd

* cleanup compiler warnings

* remove several (mostly "unused variable") compiler warnings

* remove more compiler warnings

* more compiler-warning cleanup

* int -> size_t

* remove "unused function" compiler warning

* get WIN32 right

* add back default  NITF_DEFAULT_PLUGIN_PATH #define

* fix warnings about printf() format

* fix compiler warnings about using NULL

* get rid of more compiler warnings

* fix int/int64_t conversion warnings

* __PluginRegistryInitLock is only used for _WIN32

* oops...put __PluginRegistryInitLock inside the wrong #ifdef

* remove code-analysis warnings about using a NULL pointer

* restore changes lost in previous merge

* restore externals/coda-oss to same as "master"

+++ modules/c/nitf/tests/test_fhdr_clone.c
@@ -84,7 +84,7 @@ void printHdr(nitf_FileHeader* header)
         printf("\tThe length of IMAGE subheader [%d]: %ld bytes\n",
                i, (long)len);
         printf("\tThe length of the IMAGE data: %llu bytes\n\n",
-               dataLen);
+               (long long unsigned int) dataLen);
     }
 
     NITF_TRY_GET_UINT32(header->numGraphics, &num, &error);
